📝 [#49] Docs: Button에 JSdoc 작성
